Reorganized promotions query and added additional brand info fetching
