+++ core/app/[locale]/(default)/(faceted)/category/[slug]/page.tsx
@@ -16,6 +16,8 @@ import { EmptyState } from './_components/empty-state';
 import { SubCategories } from './_components/sub-categories';
 import { getCategoryPageData } from './page-data';
 
+import { getPromotions } from '../../fetch-promotions';
+
 import { Category } from './category';
 
 interface Props {
@@ -24,33 +26,6 @@ interface Props {
     locale: string;
   }>;
   searchParams: Promise<Record<string, string | string[] | undefined>>;
-}
-
-/*
-TODO: Move to separate file...
-*/
-const storeHash = process.env.BIGCOMMERCE_STORE_HASH;
-const client = process.env.BIGCOMMERCE_ACCESS_TOKEN || '';
-const tokenRest = process.env.BIGCOMMERCE_ACCESS_TOKEN || '';
-const channelId = process.env.BIGCOMMERCE_CHANNEL_ID;
-
-export async function getPromotions() {
-  const response = await fetch(`https://api.bigcommerce.com/stores/${storeHash}/v3/promotions?channels=${channelId}&sort=priority&status=ENABLED&redemption_type=AUTOMATIC`, {
-    method: "GET",
-    credentials: "same-origin",
-    headers: {
-      "X-Auth-Client": client,
-      "X-Auth-Token": tokenRest,
-      "Content-Type": "application/json",
-      "Accept": "application/json"
-    },
-    cache: 'force-cache',
-    //next: { revalidate: 3600 }
-  });
-
-  const data = await response.json();
-
-  return data.data;
 }
 
 export async function generateMetadata({ params }: Props): Promise<Metadata> {
