+++ core/app/[locale]/(default)/(faceted)/search/page.tsx
@@ -10,36 +10,11 @@ import { redirectToCompare } from '../_actions/redirect-to-compare';
 import { fetchFacetedSearch } from '../fetch-faceted-search';
 
 import { getCompareProducts } from './page-data';
+import { getPromotions } from '../fetch-promotions';
 
 import { Breadcrumbs } from '~/components/breadcrumbs';
 import { Search } from './search';
 
-/*
-TODO: Move to separate file...
-*/
-const storeHash = process.env.BIGCOMMERCE_STORE_HASH;
-const client = process.env.BIGCOMMERCE_ACCESS_TOKEN || '';
-const tokenRest = process.env.BIGCOMMERCE_ACCESS_TOKEN || '';
-const channelId = process.env.BIGCOMMERCE_CHANNEL_ID;
-
-export async function getPromotions() {
-  const response = await fetch(`https://api.bigcommerce.com/stores/${storeHash}/v3/promotions?channels=${channelId}&sort=priority&status=ENABLED&redemption_type=AUTOMATIC`, {
-    method: "GET",
-    credentials: "same-origin",
-    headers: {
-      "X-Auth-Client": client,
-      "X-Auth-Token": tokenRest,
-      "Content-Type": "application/json",
-      "Accept": "application/json"
-    },
-    cache: 'force-cache',
-    //next: { revalidate: 3600 }
-  });
-
-  const data = await response.json();
-
-  return data.data;
-}
 
 export async function generateMetadata() {
   const t = await getTranslations('Search');
